// ui/packages/src/api/generated/filesys/schemas/createShareCommand.ts
/**
 * Generated by orval v6.31.0 🍺
 * Do not edit manually.
 * Refactored free-fs with a subject-match-java style layered architecture.
 * OpenAPI spec version: v1.0.0
 */

export interface CreateShareCommand {
  /** No comments found. */
  expireTime?: string;
  /** No comments found. */
  expireType?: number;
  /** No comments found. */
  fileIds: string[];
  /** No comments found. */
  maxDownloadCount?: number;
  /** No comments found. */
  maxViewCount?: number;
  /** No comments found. */
  needShareCode?: boolean;
  /** No comments found. */
  scope?: string;
  /** No comments found. */
  shareName?: string;
  /** No comments found. */
  userId: string;
}